refactor + fix
