+++ src/main/java/faang/school/analytics/listener/PremiumEventListener.java
@@ -1,20 +1,23 @@
 package faang.school.analytics.listener;
 
 import faang.school.analytics.dto.premium.PremiumEvent;
-import faang.school.analytics.service.PremiumEventService;
+import faang.school.analytics.mapper.AnalyticsEventMapper;
+import faang.school.analytics.service.AnalyticsEventService;
 import org.springframework.stereotype.Component;
 
 @Component
 public class PremiumEventListener extends AbstractEventListener<PremiumEvent> {
-    private final PremiumEventService premiumEventService;
+    private final AnalyticsEventMapper analyticsEventMapper;
+    private final AnalyticsEventService analyticsEventService;
 
-    public PremiumEventListener(PremiumEventService premiumEventService) {
+    public PremiumEventListener(AnalyticsEventMapper analyticsEventMapper, AnalyticsEventService analyticsEventService) {
         super(PremiumEvent.class);
-        this.premiumEventService = premiumEventService;
+        this.analyticsEventMapper = analyticsEventMapper;
+        this.analyticsEventService = analyticsEventService;
     }
 
     @Override
     protected void processEvent(PremiumEvent event) {
-        premiumEventService.save(event);
+        analyticsEventService.saveEvent(analyticsEventMapper.toEntity(event));
     }
 }
